Remove save tutorial sections
