+++ Tutorial.md
@@ -245,34 +245,6 @@
 ```
 
 ```
-このゲームはオートセーブに対応しており、
-ログや住人たちの状態は、常に自動で保存されています。
-
-ただし、ブラウザを閉じたり、キャッシュを消去したりすると、
-これらのデータは失われてしまうため注意が必要です。
-
-「セーブ」からは、セーブデータをファイルとしてダウンロードできます。
-ゲームを閉じる前には、必ずセーブを行ってデータを保存してください。
-
-また、いつでも「ロード」からファイルを読み込むことで、
-保存時点からプレイを再開することができます。
-```
-
-```
-「リセット」を押すと、現在のデータが初期化され、
-最初のセーブデータ選択に戻ります。
-
-チュートリアルも、最初からやり直すことができます。
-
-なお、事前に「セーブ」しておいたファイルがあれば、
-「ロード」からデータを復元することも可能です。
-```
-
-```
-今の状態をセーブしておきましょう。
-```
-
-```
 画面右上には、現在の日時が表示されています。
 
 ゲーム内の時間は、現実世界の時間と連動しており、
@@ -340,5 +312,4 @@
 これで、あなたはもう立派な管理人です。
 住人たちの毎日を、これからゆっくりと見守ってあげてくださいね。
 
-そして、こまめなセーブもお忘れなく。
 ```
