+++ client/src/components/SettingsPage.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useState, useRef } from 'react'
 import { useNavigate } from 'react-router-dom'
 import {
   GoogleAuthProvider,
@@ -9,11 +9,12 @@ import {
 } from 'firebase/auth'
 import { auth } from '../firebaseConfig.js'
 
-export default function SettingsPage() {
+export default function SettingsPage({ onSave, onLoad, onReset }) {
   const navigate = useNavigate()
   const [user, setUser] = useState(auth.currentUser)
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
+  const fileInputRef = useRef(null)
 
   useEffect(() => {
     const unsub = onAuthStateChanged(auth, u => setUser(u))
@@ -80,26 +81,46 @@ export default function SettingsPage() {
   return (
     <section className="p-2">
       <h1 className="text-xl font-bold mb-4">設定</h1>
+      <h2 className="text-lg mb-2">セーブデータ</h2>
+      <div className="flex flex-col gap-2 mb-4 w-fit">
+        <button onClick={onSave}>セーブ</button>
+        <button onClick={() => fileInputRef.current?.click()}>ロード</button>
+        <button onClick={onReset}>リセット</button>
+        <input
+          type="file"
+          accept="application/json"
+          ref={fileInputRef}
+          className="hidden"
+          onChange={e => {
+            const file = e.target.files?.[0]
+            if (file) onLoad(file)
+            e.target.value = ''
+          }}
+        />
+      </div>
       <h2 className="text-lg mb-2">アカウント連携</h2>
-      <div className="flex flex-col gap-2">
+      <div className="flex flex-col gap-2 w-fit">
         {user.isAnonymous && !linkedGoogle && (
           <button onClick={handleLinkGoogle}>Googleアカウントと連携する</button>
         )}
         {linkedGoogle && (
-          <button onClick={handleUnlinkGoogle}>Googleアカウントの連携を解除する</button>
+          <div className="flex flex-col gap-1">
+            <p>Googleアカウントと連携済みです</p>
+            <button onClick={handleUnlinkGoogle}>連携を解除する</button>
+          </div>
         )}
         {user.isAnonymous && !linkedEmail && (
-          <div className="flex flex-col gap-2">
+          <div className="flex flex-col gap-2 w-fit">
             <input
               type="email"
-              className="text-black p-1"
+              className="text-black p-1 w-60"
               placeholder="メールアドレス"
               value={email}
               onChange={e => setEmail(e.target.value)}
             />
             <input
               type="password"
-              className="text-black p-1"
+              className="text-black p-1 w-60"
               placeholder="パスワード"
               value={password}
               onChange={e => setPassword(e.target.value)}
@@ -108,7 +129,10 @@ export default function SettingsPage() {
           </div>
         )}
         {linkedEmail && (
-          <button onClick={handleUnlinkEmail}>メールアドレス連携を解除する</button>
+          <div className="flex flex-col gap-1">
+            <p>メールアドレスと連携済みです</p>
+            <button onClick={handleUnlinkEmail}>連携を解除する</button>
+          </div>
         )}
       </div>
       <button className="mt-4" onClick={() => navigate(-1)}>戻る</button>
